requirements.txt
hydra-core==1.2.0
numpy==1.22.3
omegaconf==2.2.2
pkbar==0.5
wget==3.2
